I'm abandoning this branch of the code.
I need to rework the way I'm doing this.  I'm really unhappy with the
way that I laid out this program when I started and now it is very
difficult to change anything without destroying the whole program.  So
it is time to start fresh.

This is getting committed for reference purposes and nothing more.