Added a told reps or reps program thinks you should do to db layout.
This is because it needs more info than what you did.  It also needs to
know what you intended to do, which is most often what this program told
you to attempt.  That way it can better determine what the hell you
intended on doing doing when your weights and reps go up and down randomly when you mess things up.