Untested, but I feel like this might be a slight optimization.
When weights are put on they are always in descending order, but
when plate orders are chosen, they are always in ascending if tied,
which is the exact opposite of what I want.  I have a feeling
that by putting this reverse in before I run minimumBy, the function
will pick a slightly more preferable plate order when several have
the same scores otherwise.